rename: global response dir -> dto로 수정
