+++ src/main/java/com/example/techit7/comment/dto/CommentResponseDto.java
@@ -1,7 +1,7 @@
 package com.example.techit7.comment.dto;
 
 import com.example.techit7.comment.entity.Comment;
-import com.example.techit7.global.response.GlobalResponseDto;
+import com.example.techit7.global.dto.GlobalResponseDto;
 
 public class CommentResponseDto extends GlobalResponseDto<Comment> {
