+++ src/main/java/com/example/techit7/comment/service/CommentService.java
@@ -2,12 +2,10 @@ package com.example.techit7.comment.service;
 
 import com.example.techit7.article.entity.Article;
 import com.example.techit7.comment.dto.CommentRequestDto;
-import com.example.techit7.comment.dto.CommentResponseDto;
 import com.example.techit7.comment.entity.Comment;
 
-import com.example.techit7.global.response.GlobalResponseDto;
+import com.example.techit7.global.dto.GlobalResponseDto;
 import com.example.techit7.user.User;
-import java.util.List;
 
 public interface CommentService {
